+++ src/main/java/com/algaworks/algafood/infraestructore/repository/EstadoRepositoryImpl.java
@@ -2,6 +2,7 @@ package com.algaworks.algafood.infraestructore.repository;
 
 import com.algaworks.algafood.domain.model.Estado;
 import com.algaworks.algafood.domain.repository.EstadoRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,9 +35,11 @@ public class EstadoRepositoryImpl implements EstadoRepository {
 
     @Transactional
     @Override
-    public void remover(Estado estado) {
+    public void remover(Long id) {
+        Estado estado = buscarPorId(id);
 
-        estado = buscarPorId(estado.getId());
+        if(estado == null)
+            throw new EmptyResultDataAccessException(1);
 
         em.remove(estado);
     }
